% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_color.R
\name{add_color}
\alias{add_color}
\title{Add vector of ICPI colors}
\usage{
add_color(palette = "woods")
}
\arguments{
\item{palette}{which color palette to pull in (lower case, last word), Default = "woods"}
}
\description{
The ICPI color palettes can be imported as hex vectors into R. To do so, you can use the `add_colors()` function to extract from Github.

There are four color palette. To simplify the input, you will just use the last word of the color palette name in lower case.

Palette Name | Input
-- | --
  Autumn Woods | "woods"
Coast of Bohemia | "bohemia"
Tidepools | "tidepools"
By the Power of Grayscale | "grayscale"
}
\examples{
\dontrun{
#pull in Autumn Woods hex colors as vector
  palette_woods <- add_color()
#pull in By the Power of Grayscale colors as vector
  palette_gray <- add_color("grayscale") }

}
