% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_cumulative.R
\name{add_cumulative}
\alias{add_cumulative}
\title{Add cumulative value for current FY}
\usage{
add_cumulative(df)
}
\arguments{
\item{df}{data frame to add cumulative column onto}
}
\description{
This function is adapted from the achafetz/PartnerProgress repo
}
\examples{
\dontrun{
df_mer <- add_cumulative(df_mer)}
}
