% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adj_ovc_apr17.R
\name{adj_ovc_apr17}
\alias{adj_ovc_apr17}
\title{Add FY17 APR OVC_SERV Total Numerator}
\usage{
adj_ovc_apr17(df)
}
\arguments{
\item{df}{dataframe to adjust OVC_SERV FY17 APR targets}
}
\description{
Add FY17 APR OVC_SERV Total Numerator
}
