% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_achievement.R
\name{adorn_achievement}
\alias{adorn_achievement}
\title{Adorn Achievement}
\usage{
adorn_achievement(df, qtr = NULL)
}
\arguments{
\item{df}{data frame as standrd MSD or one from reshape_msd()}

\item{qtr}{if using standard MSD, need to provide the most recent quarter,
ideally using identifypd(df_msd, pd_type = "quarter")}
}
\value{
data frame with achievement values, labels, and colors
}
\description{
Function to calculate target achievement (cumulative and/or quarterly) for a
standard MSD or one reshaped using reshape_msd() as well as to apply achievement
group labels and colors.
}
\examples{
\dontrun{
df_msd <- read_msd(path)
df_msd_agg <- df_msd \%>\%
 filter(operatingunit == "Jupiter"
        indicator \%in\% c("TX_NEW", "TX_CURR"),
        fundingagency != "Dedup",
        standardizeddisaggregate == "Total Numerator") \%>\%
 group_by(operatingunit, fundingagency, fiscal_year, indicator) \%>\%
 summarise(across(where(is.double), sum, na.rm = TRUE)) \%>\%
 ungroup()

adorn_achievement(df_msd_agg)

df_msd_agg \%>\%
 reshape_msd("quarters") \%>\%
 adorn_achievement()

df_msd_agg \%>\%
 reshape_msd("quarters", qtrs_keep_cumulative = TRUE) \%>\%
 adorn_achievement() }
}
