% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_achievement.R
\name{assign_quarter}
\alias{assign_quarter}
\title{Assign Quarter}
\usage{
assign_quarter(df, curr_qtr = 4)
}
\arguments{
\item{df}{MSD dataframe, standard or reshaped via reshape_msd()}

\item{curr_qtr}{quarter for current fiscal year (if standard MSD)}
}
\value{
data frame with a quarter column
}
\description{
Assign Quarter
}
