% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_achievement.R
\name{calc_achievement}
\alias{calc_achievement}
\title{Calculate Achievement}
\usage{
calc_achievement(df)
}
\arguments{
\item{df}{MSD based dataframe}
}
\value{
one or two additional columns that calculate achievement and/or quarterly achievement
}
\description{
Calculate Achievement
}
\examples{
\dontrun{
df_msd <- read_msd(path)
df_msd_agg <- df_msd \%>\%
 filter(operatingunit == "Jupiter"
        indicator \%in\% c("TX_NEW", "TX_CURR"),
        fundingagency != "Dedup",
        standardizeddisaggregate == "Total Numerator") \%>\%
 group_by(operatingunit, fundingagency, fiscal_year, indicator) \%>\%
 summarise(dplyr::across(where(is.double), sum, na.rm = TRUE)) \%>\%
 ungroup()

calc_achievement(df_msd_agg)

df_msd_agg \%>\%
 reshape_msd("quarters") \%>\%
 calc_achievement() }
}
