% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_genpop.R
\name{calc_genpop}
\alias{calc_genpop}
\title{Calculate General Population PEPFAR results}
\usage{
calc_genpop(df)
}
\arguments{
\item{df}{PEPFAR MER structured dataset in data.frame format for calculation, such as the object created by ICPIutilities::read_msd()}
}
\value{
a simplified summarised dataframe with a standardized kpgroup variable and results for general population
}
\description{
Standardizes key populations groupings and calculates estimated general population result
by subtracting key populations results from Total Numerator
}
