% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_achievement.R
\name{color_achievement}
\alias{color_achievement}
\title{Color and label achievement}
\usage{
color_achievement(df, curr_qtr = NULL)
}
\arguments{
\item{df}{MSD dataframe}

\item{curr_qtr}{if wide MSD, need to specify the curren period, defaults to Q4}

\item{input_type}{whether to handle 'achievement' or 'achievement_qtrly'}
}
\value{
dataframe with achievement colors and labels
}
\description{
Color and label achievement
}
