% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_netnew.R
\name{combine_netnew}
\alias{combine_netnew}
\title{Add NET_NEW to dataset (orginally in achafetz/PartnerProgress)}
\usage{
combine_netnew(df)
}
\arguments{
\item{df}{dataframe to create net new based off of and append onto}
}
\description{
Add NET_NEW to dataset (orginally in achafetz/PartnerProgress)
}
\examples{
\dontrun{
df_msd <- combine_netnew(df_msd)}


}
