% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dp_compair.R
\name{compair}
\alias{compair}
\title{compare PSNUxIM pairs in Original v Current DP allocation tab}
\usage{
compair(new_file, orig_file, folderpath)
}
\arguments{
\item{new_file}{file name of the new Data Pack including .xlsx}

\item{orig_file}{file name of the original Data Pack including .xlsx}

\item{folderpath}{folder location where both data pack's are stored}
}
\value{
PSNU, IM and Type columns comparison to use in Disagg Tools
}
\description{
compare PSNUxIM pairs in Original v Current DP allocation tab
}
