% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_pairs.R
\name{compare_pairs}
\alias{compare_pairs}
\title{compare PSNUxIM pairs in Original v Current DP allocation tab}
\usage{
compare_pairs(new_file, orig_file, folderpath)
}
\arguments{
\item{new_file}{file name of the new Data Pack including .xlsx}

\item{orig_file}{file name of the original Data Pack including .xlsx}

\item{folderpath}{folder location where both data pack's are stored}
}
\value{
PSNU, IM and Type columns comparison to use in Disagg Tools
}
\description{
compare PSNUxIM pairs in Original v Current DP allocation tab
}
