% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dp_allocissues.R
\name{flag_error}
\alias{flag_error}
\title{Identify Neg or Missing Errors with HTS, TX, OVC, and VMMC in Allocation Tab}
\usage{
flag_error(filepath, file)
}
\arguments{
\item{filepath}{where is the folder that contains the Data Pack or Disagg Tool}

\item{file}{what is the file name of the DP/DT}
}
\value{
lines where there is a numerator/disagg discrepency
}
\description{
Identify Neg or Missing Errors with HTS, TX, OVC, and VMMC in Allocation Tab
}
