% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_netnew.R
\name{gen_netnew}
\alias{gen_netnew}
\title{Create Net New Variable}
\usage{
gen_netnew(df, type = "result")
}
\arguments{
\item{df}{data frame to use}

\item{type}{either result or target, default = result}
}
\description{
Create Net New Variable
}
\examples{
\dontrun{
df_mds_results <- gen_netnew(df_mds, type = "result")}
}
