% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dp_allocissues.R
\name{identify_dups}
\alias{identify_dups}
\title{Identify duplicates}
\usage{
identify_dups(filepath, file)
}
\arguments{
\item{filepath}{where is the folder that contains the Data Pack or Disagg Tool}

\item{file}{what is the file name of the DP/DT}
}
\value{
duplicates count (PSNU-MechanismID-Type combos should be unique)
}
\description{
Identify duplicates
}
