% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identifypd.R
\name{identifypd}
\alias{identifypd}
\title{Extract lastest period of dataset}
\usage{
identifypd(df, pd_type = "full")
}
\arguments{
\item{df}{dataset to use to find latest period}

\item{pd_type}{what is returned? (a) full, eg fy2018q1; (b)year, eg 2018; (c) or quarter, eg 1}
}
\description{
This function is pulled from the achafetz/PartnerProgress repo
}
\examples{
\dontrun{
identifypd(df_mer)
identifypd(df_mer, "quarter") }

}
