% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alloctab_issues.R
\name{import}
\alias{import}
\title{IMPORT DATA PACK OR DISAGG TOOL}
\usage{
import(filepath, file)
}
\arguments{
\item{filepath}{where is the folder that contains the Data Pack or Disagg Tool}

\item{file}{what is the file name of the DP/DT}
}
\value{
import datafram
}
\description{
IMPORT DATA PACK OR DISAGG TOOL
}
