% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_pairs.R
\name{import_pairs}
\alias{import_pairs}
\title{import PSNUxIM pairs from DP Allocation tab}
\usage{
import_pairs(folderpath, filename, type)
}
\arguments{
\item{folderpath}{folder location where the data pack is stored}

\item{filename}{file name of the data pack including .xlsx}

\item{type}{either orig or new}
}
\value{
PSNU, IM and Type columns from the DP
}
\description{
import PSNUxIM pairs from DP Allocation tab
}
