% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_msd.R
\name{match_msd}
\alias{match_msd}
\title{Transmute DATIM Genie output to match MER Structured Dataset}
\usage{
match_msd(genie_filepath)
}
\arguments{
\item{genie_filepath}{file path to the zipped DATIM Genie file}
}
\description{
Transmute DATIM Genie output to match MER Structured Dataset
}
\examples{
\dontrun{
df_genie <- match_msd("~/Downloads/PEPFAR-Data-Genie-PSNUByIMs-2018-08-15.zip") }

}
