% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_fv.R
\name{read_fv}
\alias{read_fv}
\title{Import Fact View Dataset .txt into R and covert to .rds}
\usage{
read_fv(file, path = "~/ICPI/Data/", save_rds = TRUE, remove_txt = FALSE)
}
\arguments{
\item{file}{What is the file name, eg ICPI_FactView_PSNU_20171222_v2_1.txt?}

\item{path}{file path to the parent folder, default = "~/ICPI/Data/"}

\item{save_rds}{save the Fact View as an Rds file, default = TRUE}

\item{remove_txt}{should the txt file be removed, default = FALSE}
}
\description{
This funciton imports a stored ICPI Fact View Dataset and coverts it from a .txt to an .Rds to significantly limit file size
}
\examples{

#convert Q4 clean OUxIM file from txt to Rds
 df_fv_psnu <- read_fv("ICPI_FactView_PSNU_20171222_v2_1.txt", path = "~/Downloads/")
#import Q1 PSNU
 df_pnsu <-  read_fv("ICPI_FactView_PSNU_20180215_v1_3.txt", path = "~/Downloads/", save_rds = FALSE)

}
