% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_msd.R
\name{read_msd}
\alias{read_msd}
\title{Import ICPI MER Structured Datasets .txt into R and covert to .rds}
\usage{
read_msd(file, save_rds = FALSE, remove_txt = FALSE)
}
\arguments{
\item{file}{enter the full path to the MSD/FSD file}

\item{save_rds}{save the Structured Dataset as an rds file, default = FALSE}

\item{remove_txt}{should the txt file be removed, default = FALSE}
}
\description{
This function imports a stored ICPI MER/ER Structured Datasets and coverts it from a .txt to an .Rds to significantly limit file size
}
\examples{

\dontrun{#convert Q1 clean PSNU file from txt to Rds
#read in file for use
 path <- "~/Data/ICPI_MER_Structured_Dataset_PSNU_20180323_v2_1.txt"
 df_psnu <- read_msd(path)
#convert to RDS and delete the original txt file
 read_msd(path, save_rds = TRUE, remove_txt = TRUE)}

}
