% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_msd.R
\name{rename_msd}
\alias{rename_msd}
\title{Rename MSD file when importing}
\usage{
rename_msd(file)
}
\arguments{
\item{file}{enter the full path to the MSD/ERSD file, eg "~/ICPI/Data/ICPI_MER_Structured_Dataset_PSNU_20180323_v2_1.txt"}
}
\description{
Rename MSD file when importing
}
