% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reshape_msd.R
\name{reshape_msd}
\alias{reshape_msd}
\title{Reshape Semi-Wide MSD}
\usage{
reshape_msd(df, direction = c("wide", "long"))
}
\arguments{
\item{df}{MSD dataset in the semi-wide format}

\item{direction}{direction of reshape, wide (default) or long}
}
\description{
Reshape Semi-Wide MSD
}
\examples{
\dontrun{
 #read in data
  df_genie <- match_msd("~/Downloads/PEPFAR-Data-Genie-PSNUByIMs-2018-08-15.zip")
 #reshape long
  df_genie_long <- reshape_msd(df_genie, direction = "long")
 #or reshape wide
  df_genie_wide <- reshape_msd(df_genie, direction = "wide") }
}
