% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_save.R
\name{split_save}
\alias{split_save}
\title{Export a single dataset into multiple files by group}
\usage{
split_save(df, folderpath, filename_stub)
}
\arguments{
\item{df}{dataframe to split}

\item{folderpath}{directory where you want to store the files}

\item{filename_stub}{generic stub for naming all the files}
}
\description{
Export a single dataset into multiple files by group
}
\examples{
\dontrun{
  split_save(df, "~/CountryFiles", "FY18Q3_TX")
}
}
