% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_msd.R
\name{unzip_msd}
\alias{unzip_msd}
\title{Unzip packaged MSD}
\usage{
unzip_msd(msdfilepath_zip, remove_zip = FALSE)
}
\arguments{
\item{msdfilepath_zip}{full file path of zipped MSD}

\item{remove_zip}{after extracting the flat file, do you want the zipped folder removed?}
}
\description{
Unzip packaged MSD
}
