% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{var_exists}
\alias{var_exists}
\title{Check if all variables in list exist}
\usage{
var_exists(df, vars)
}
\arguments{
\item{df}{data frame}

\item{vars}{quoted variable(s)}
}
\description{
Check if all variables in list exist
}
