% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{var_missing}
\alias{var_missing}
\title{Check if all variables in list are missing}
\usage{
var_missing(df, vars)
}
\arguments{
\item{df}{data frame}

\item{vars}{quoted variable(s)}
}
\description{
Check if all variables in list are missing
}
