% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_msd.R
\name{match_msd}
\alias{match_msd}
\title{Transmute DATIM Genie output to match MER Structured Dataset}
\usage{
match_msd(genie_filepath, to_lower = TRUE, save_rds = TRUE)
}
\arguments{
\item{genie_filepath}{enter the full file path to the zipped DATIM Genie file}

\item{to_lower}{do you want to convert all names to lower case, default = TRUE}

\item{save_rds}{save the Structured Dataset as an Rds file, default = TRUE}
}
\description{
Transmute DATIM Genie output to match MER Structured Dataset
}
\examples{
\dontrun{
df_genie <- match_msd("~/Downloads/PEPFAR-Data-Genie-PSNUByIMs-2018-08-15.zip") }

}
