% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_msd.R
\name{read_msd}
\alias{read_msd}
\title{Import ICPI MER Structured Datasets .txt into R and covert to .rds}
\usage{
read_msd(file, path = "~/ICPI/Data/", to_lower = TRUE, save_rds = TRUE,
  remove_txt = FALSE)
}
\arguments{
\item{file}{What is the file name, eg ICPI_MER_Structured_Dataset_PSNU_20180323_v2_1.txt?}

\item{path}{file path to the parent folder, default = "~/ICPI/Data/"}

\item{to_lower}{do you want to convert all names to lower case, default = TRUE}

\item{save_rds}{save the Structured Dataset as an Rds file, default = TRUE}

\item{remove_txt}{should the txt file be removed, default = FALSE}
}
\description{
This funciton imports a stored ICPI MER Structured Datasets and coverts it from a .txt to an .Rds to significantly limit file size
}
\examples{

\dontrun{#convert Q4 clean OUxIM file from txt to Rds
 df_psnu <- read_msd("ICPI_MER_Structured_Dataset_PSNU_20180323_v2_1.txt",
 path = "~/Downloads/")
#import Q1 PSNU
 df_pnsu <-  read_msd("ICPI_MER_Structured_Dataset_PSNU_20180323_v2_1.txt",
 path = "~/Downloads/", save_rds = FALSE)}

}
