% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_msd.R
\name{read_msd}
\alias{read_msd}
\title{Import ICPI MER Structured Datasets .txt into R and covert to .rds}
\usage{
read_msd(file, to_lower = TRUE, save_rds = TRUE, remove_txt = FALSE)
}
\arguments{
\item{file}{enter the full path to the MSD file, eg "~/ICPI/Data/ICPI_MER_Structured_Dataset_PSNU_20180323_v2_1.txt"}

\item{to_lower}{do you want to convert all names to lower case, default = TRUE}

\item{save_rds}{save the Structured Dataset as an Rds file, default = TRUE}

\item{remove_txt}{should the txt file be removed, default = FALSE}
}
\description{
This funciton imports a stored ICPI MER Structured Datasets and coverts it from a .txt to an .Rds to significantly limit file size
}
\examples{

\dontrun{#convert Q1 clean PSNU file from txt to Rds
#read in file for use (without saving as an RDS)
   df_psnu <- read_msd("~/ICPI/Data/ICPI_MER_Structured_Dataset_PSNU_20180323_v2_1.txt", save_rds = FALSE)
#convert to RDS and delete the original txt file
 read_msd("~/ICPI/Data/ICPI_MER_Structured_Dataset_PSNU_20180323_v2_1.txt", remove_txt = TRUE)}

}
