% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_save.R
\name{split_save}
\alias{split_save}
\title{Export a single dataset into multiple files by group}
\usage{
split_save(df, group_var, folderpath, filename_stub, include_date = FALSE)
}
\arguments{
\item{df}{dataframe to split}

\item{group_var}{grouping variable to split the dataset by, eg operatingunit, fundingagency}

\item{folderpath}{directory where you want to store the files}

\item{filename_stub}{generic stub for naming all the files}

\item{include_date}{include date after filenamestub? default = FALSE, eg "20180913"}
}
\description{
Export a single dataset into multiple files by group
}
\examples{
\dontrun{
#create country specific files for TX_NEW
 df_mer \%>\%
  filter(indicator == "TX_NEW",
         standardizeddisaggregate == "Total Numerator") \%>\%
  split_save(operatingunit, "~/CountryFiles", "FY18Q3_TX")
}
}
